Extract common row path traversal for Grid and Boustrophedon in a separate module.
