+++ src/libslic3r/GCode/GridPath.cpp
@@ -3,21 +3,14 @@
 
 #include "GridPath.hpp"
 #include "TSPPostProcessing.hpp"
-#include "../Print.hpp"
 #include "../Geometry.hpp"
 
-#include <algorithm>
 #include <cmath>
 #include <limits>
-#include <unordered_map>
 #include <utility>
 #include <vector>
 
 namespace Slic3r {
-
-// Row threshold parameters (in µm).
-constexpr double ROW_FRACTION_OF_Y_RANGE = 0.1;   // ~10% of Y range
-constexpr double MIN_ROW_THRESHOLD_UM     = 1e4;   // 10 mm floor to avoid over-splitting
 
 // Manhattan distance between two points.
 static double manhattan_dist(const Point& a, const Point& b)
@@ -136,59 +129,8 @@ std::vector<size_t> grid_path_core(const Points& centers)
 {
     if (centers.empty()) return {};
 
-    size_t n = centers.size();
-
-    // Compute Y range to determine row grouping threshold.
-    double y_min = std::numeric_limits<double>::max();
-    double y_max = -std::numeric_limits<double>::max();
-    for (const auto& p : centers) {
-        double y = static_cast<double>(p.y());
-        if (y < y_min) y_min = y;
-        if (y > y_max) y_max = y;
-    }
-
-    // Row threshold: fraction of Y range, with a minimum floor.
-    double row_threshold = (y_max - y_min) * ROW_FRACTION_OF_Y_RANGE;
-    if (row_threshold < MIN_ROW_THRESHOLD_UM) row_threshold = MIN_ROW_THRESHOLD_UM;
-
-    // Group into rows using hash-map binning by Y coordinate.
-    std::unordered_map<int, std::vector<size_t>> row_map;
-    for (size_t i = 0; i < n; ++i) {
-        int y_key = static_cast<int>(centers[i].y() / row_threshold);
-        row_map[y_key].push_back(i);
-    }
-
-    // Convert to vector of rows with precomputed average Y.
-    struct Row { double avg_y; std::vector<size_t> indices; };
-    std::vector<Row> rows;
-    rows.reserve(row_map.size());
-    for (auto& [key, indices] : row_map) {
-        double sum = 0;
-        for (size_t idx : indices) sum += static_cast<double>(centers[idx].y());
-        rows.push_back({sum / static_cast<double>(indices.size()), std::move(indices)});
-    }
-
-    // Sort rows by average Y.
-    std::sort(rows.begin(), rows.end(), [](const auto& a, const auto& b) {
-        return a.avg_y < b.avg_y;
-    });
-
-    // Sort each row by X coordinate, then traverse alternating direction (serpentine).
-    std::vector<size_t> path;
-    path.reserve(n);
-    bool reverse = false;
-
-    for (auto& [avg_y, row] : rows) {
-        std::sort(row.begin(), row.end(),
-            [&](size_t a, size_t b) { return centers[a].x() < centers[b].x(); });
-
-        if (reverse) {
-            path.insert(path.end(), row.rbegin(), row.rend());
-        } else {
-            path.insert(path.end(), row.begin(), row.end());
-        }
-        reverse = !reverse;
-    }
+    // Row-based serpentine traversal.
+    std::vector<size_t> path = row_serpentine_path(centers);
 
     // L1-aware 2-opt: only accepts swaps that improve Manhattan distance.
     tsp_2opt_l1_improve(path, centers);
